Move BackupLoaderService to infrastructure
